+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 0.51.0
 
 ### Enhancements
+- **Event Changes**
+  - Event completion percentage is now based on total number of questions and
+    responses of the event's sheets
 - **Sheet Changes**
   - Improved the load time of sheet show pages
   - Improved the load time of sheet edit pages